+++ CHANGELOG.md
@@ -86,4 +86,4 @@
 
 ### Features
 
-* full FraudIntelligence platform (docs, infra, services, ml, slides) ([5cb7adc](https://github.com/JRmon42/FraudIntelligence/commit/5cb7adc70cdfc6c0067bd6b06e54558270e513fa))
+* full Heimdall platform (docs, infra, services, ml, slides) ([5cb7adc](https://github.com/JRmon42/FraudIntelligence/commit/5cb7adc70cdfc6c0067bd6b06e54558270e513fa))